refactor: change tag validation name
