+++ src/application/validation/validators/index.ts
@@ -21,4 +21,4 @@ export * from './days-of-week-validation'
 export * from './days-of-month-validation'
 export * from './daily-validation'
 export * from './color-validation'
-export * from './tagId-validator-adapter'
+export * from './tagId-validation'
